pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "jobFinder"
version = "0.0.1"
dependencies = [
    "annotated-types",
    "anyio",
    "asttokens",
    "Authlib",
    "blinker",
    "certifi",
    "cffi",
    "charset-normalizer",
    "click",
    "cryptography",
    "decorator",
    "distro",
    "executing",
    "Flask",
    "grpcio",
    "grpcio-health-checking",
    "grpcio-tools",
    "h11",
    "httpcore",
    "httpx",
    "idna",
    "itsdangerous",
    "jedi",
    "Jinja2",
    "joblib",
    "MarkupSafe",
    "matplotlib-inline",
    "nltk",
    "openai",
    "parso",
    "pexpect",
    "prompt_toolkit",
    "protobuf",
    "ptyprocess",
    "pure-eval",
    "pycparser",
    "pydantic",
    "pydantic_core",
    "Pygments",
    "python-dotenv",
    "regex",
    "requests",
    "setuptools",
    "six",
    "sniffio",
    "stack-data",
    "tqdm",
    "traitlets",
    "typing_extensions",
    "urllib3",
    "validators",
    "wcwidth",
    "weaviate-client",
    "Werkzeug"
]